rtmpt: Fix writing HTTP responses to packet body.
Now done first thing when the body is allocated. This implies
the size passed into the body must be exact for Content-Length
to be accurate.

A bit of a kludge: involves manually disabling the poll
interval in the cases it doesn't need to be written. See
HTTP_respond for this.