rtmpt: Yet another rewrite of HTTP sending.
1. Keep headers in a separate buffer from the RTMP packet body
2. Track header/body transmission as discrete messages

Also requires some extra logic in eg, AllocateWB to keep track of
buffer resets and reallocs, so the pointers for RTMPT messages can
be adjusted accordingly.

Still needs work; already-queued but partially written messages
are NOT adjusted for reallocs! Must be done!

The pollInterval is also now used to determine whether the client
is expecting a response; we don't write out packets before then.