server: Refactor initial play messages.
Makes it easier to resend the whole batch if a
waiting stream starts publishing, for example.

Information from the previous packets are not really
needed; txn id should be 0 since we're not expecting a
response from the status messages, and the internal
streamid should be the same as the packet streamid.

That is to say, however, that multiple streams/streamids per
connection are not currently supported.